==> users/ASA24 Do not edit/25_k-means_ASA24.R <==
# ===============================================================================================================
# k-means clustering with ASA24 data.
# Version 1
# Created on 01/06/2022 by Rie Sadohara
# ===============================================================================================================

# Set your working directory as to the main directory.
  Session --> Set working direHctory --> Choose directory.

# Name your main directory for future use. 
  main_wd <- file.path(getwd())

# Import source code to run the analyses to follow.
  source("lib/specify_data_dir.R")
  source("lib/k-means.R")
  
# You can come back to the main directory by:
  setwd(main_wd) 
  
# ===============================================================================================================
# Import data from your data directory 
# ===============================================================================================================

# Specify the directory where the data is.
  SpecifyDataDirectory(directory.name = "eg_data/VVKAJ/")

# Here, we will run k-means analysis for each of Nutrient as is, Nutrient average, Food categories as is, 
# and Food categories average.
  
# ===============================================================================================================
# Nutrient data as is, processed for clustering analyses.
# ===============================================================================================================

# Load Nut_asis data.
  Tot_m_QCed_Nut_asis <- read.table(file="VVKAJ_Tot_m_QCed_Nut_asis.txt", sep="\t", header=T)
  
# Scale your input file and name it as k-means_input.
  kmeans_input <- scale(Tot_m_QCed_Nut_asis)
  
# Ensure your input file has the correct number of rows and columns.
  dim(kmeans_input)
  
# Specify the directory (folder) to save the results.
  res_dir_nut_asis = "kmeans_Nut_asis"
  
# Specify the prefix of filenames to be saved. 
  res_prefix_nut_asis = "VVKAJ_Nut_asis"
  
# Run elbow, silhouette, and gap methods to find an optimum K (number of clusters). 
# Do not alter the name of the input file: kmeans_input. This function below assumes that
# the input is named as "kmeans_input". 
# You can only run those three methods for K = 1 through (number of observations - 1). 
# The gap method output will be printed on the Console. The gap values are plotted in 
# xxx_gapmethod.pdf.
  ChooseK(out.dir= res_dir_nut_asis, out.prefix= res_prefix_nut_asis)

# Look at the three figures generated by the ChooseK function above. The elbow and gap methods did not give a 
# distinct peak, but the silhouette method gave a peak at K=3. K=2,4,5 also has relatively high silhouette 
# width (~goodness of fit).
  
# With specific K values in mind, perform k-means analysis with one specified K.
# Also, change the file name to be saved as a PDF. 
  OneK(myK= 3, out.dir= res_dir_nut_asis, out.fn = "VKAJ_Nut_asis_K3")
  
# Or try multiple Ks and print the biplots in one panel.
# Likewise, change the file name to be saved as a PDF as necessary. 
# This uses the factoextra and gridExtra packages.  
  MultipleK(myKs = c(3,4,5,6), out.dir = res_dir_nut_asis, out.fn = "VKAJ_Nut_asis_K3-6")

  
# ===============================================================================================================
# Nutrient data averaged across days, processed for clustering analyses.
# ===============================================================================================================
  
# Load Nut_ave data.
  Tot_m_QCed_Nut_ave <- read.table(file="VVKAJ_Tot_m_QCed_Nut_ave_subset.txt", sep="\t", header=T)
  
# Scale your input file and name it as k-means_input.
  kmeans_input <- scale(Tot_m_QCed_Nut_ave)
  
# Ensure your input file has the correct number of rows and columns.
  dim(kmeans_input)
  
# Specify the directory (folder) to save the results.
  res_dir_nut_ave = "kmeans_Nut_ave" 
  
# Specify the prefix of filenames to be saved. 
  res_prefix_nut_ave = "VVKAJ_Nut_ave"
  
# Run elbow, silhouette, and gap methods to find an optimum K (number of clusters). 
  ChooseK(out.dir= res_dir_nut_ave, out.prefix= res_prefix_nut_ave)
  
# With specific K values in mind, perform k-means analysis with one specified K.
  OneK(myK= 5, out.dir= res_dir_nut_ave, out.fn = "VVKAJ_Nut_ave_K5")  
  
# Try multiple Ks and print the biplots in one panel.
  MultipleK(myKs = c(3,4,5,7), out.dir = res_dir_nut_ave, out.fn = "VVKAJ_Nut_ave_K3457")
  

# ===============================================================================================================
# Food categories data as is, processed for clustering analyses.
# ===============================================================================================================
  
# Load Cat_asis data.
  Tot_m_QCed_Cat_asis <- read.table(file="VVKAJ_Tot_m_QCed_Cat_asis.txt", sep="\t", header=T)
  
# Scale your input file and name it as k-means_input.
  kmeans_input <- scale(Tot_m_QCed_Cat_asis)
  
# Ensure your input file has the correct number of rows and columns.
  dim(kmeans_input)
  
# Specify the directory (folder) to save the results.
  res_dir_cat_asis = "kmeans_Cat_asis" 
  
# Specify the prefix of filenames to be saved. 
  res_prefix_cat_asis = "VVKAJ_Cat_asis"
  
# Run elbow, silhouette, and gap methods to find an optimum K (number of clusters). 
  ChooseK(out.dir= res_dir_cat_asis, out.prefix= res_prefix_cat_asis)
  
# With specific K values in mind, perform k-means analysis with one specified K.
  OneK(myK= 3, out.dir= res_dir_cat_asis, out.fn = "VVKAJ_Cat_asis_K3")
  
# Try multiple Ks and print the biplots in one panel.
  MultipleK(myKs = c(3,4,5,6), out.dir = res_dir_cat_asis, out.fn = "VVKAJ_Cat_asis_K3-6")
  

# ===============================================================================================================
# Food categories data averaged across days, processed for clustering analyses.
# ===============================================================================================================
  
# Load Cat_ave data.
  Tot_m_QCed_Cat_ave <- read.table(file="VVKAJ_Tot_m_QCed_Cat_ave_subset.txt", sep="\t", header=T)
  
# Scale your input file and name it as k-means_input.
  kmeans_input <- scale(Tot_m_QCed_Cat_ave)
  
# Ensure your input file has the correct number of rows and columns.
  dim(kmeans_input)
  
# Specify the directory (folder) to save the results.
  res_dir_cat_ave = "kmeans_Cat_ave" 
  
# Specify the prefix of filenames to be saved. 
  res_prefix_cat_ave = "VVKAJ_Cat_ave"
  
# Run elbow, silhouette, and gap methods to find an optimum K (number of clusters). 
  ChooseK(out.dir= res_dir_cat_ave, out.prefix= res_prefix_cat_ave)
  
# With specific K values in mind, perform k-means analysis with one specified K.
  OneK(myK= 4, out.dir= res_dir_cat_ave, out.fn = "VVKAJ_Cat_ave_K4")  
  
# Try multiple Ks and print the biplots in one panel.
  MultipleK(myKs = c(2,3,4), out.dir = res_dir_cat_ave, out.fn = "VVKAJ_Cat_ave_K2-4")

  
# ---------------------------------------------------------------------------------------------------------------
# Come back to the main directory
  setwd(main_wd)
